Blog_Generation_Gemini/requirements.txt
streamlit
python-dotenv
langchain==0.3.7
langchain-core
langchain-community==0.3.7
langchain-google-genai==1.0.3